+++ rusk-wallet/src/bin/interactive/command_menu.rs
@@ -7,7 +7,12 @@
 use execution_core::transfer::data::MAX_MEMO_SIZE;
 use requestty::Question;
 use rusk_wallet::{
-    currency::Dusk, gas, Address, Wallet, MAX_FUNCTION_NAME_SIZE,
+    currency::Dusk,
+    gas::{
+        self, DEFAULT_LIMIT_CALL, DEFAULT_LIMIT_DEPLOYMENT,
+        DEFAULT_LIMIT_STAKE, DEFAULT_LIMIT_TRANSFER,
+    },
+    Address, Wallet, MAX_FUNCTION_NAME_SIZE,
 };
 
 use crate::{prompt, settings::Settings, Command, Menu, WalletFile};
@@ -37,7 +42,7 @@ enum MenuItem {
 pub(crate) fn online(
     profile_idx: u8,
     wallet: &Wallet<WalletFile>,
-    phoenix_balance: Dusk,
+    phoenix_spendable: Dusk,
     moonlight_balance: Dusk,
     settings: &Settings,
 ) -> anyhow::Result<ProfileOp> {
@@ -70,14 +75,25 @@ pub(crate) fn online(
     let res = match cmd {
         MenuItem::Transfer => {
             let rcvr = prompt::request_rcvr_addr("recipient")?;
+
             let (sender, balance) = match &rcvr {
                 Address::Shielded(_) => {
-                    (wallet.shielded_address(profile_idx)?, phoenix_balance)
+                    (wallet.shielded_address(profile_idx)?, phoenix_spendable)
                 }
                 Address::Public(_) => {
                     (wallet.public_address(profile_idx)?, moonlight_balance)
                 }
             };
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_TRANSFER,
+                "a transfer transaction",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
 
             let memo = Some(prompt::request_str("memo", MAX_MEMO_SIZE)?);
             let amt = if memo.is_some() {
@@ -98,14 +114,23 @@ pub(crate) fn online(
             }))
         }
         MenuItem::Stake => {
-            let (addr, balance) = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    (wallet.shielded_address(profile_idx)?, phoenix_balance)
-                }
-                prompt::TransactionModel::Public => {
-                    (wallet.public_address(profile_idx)?, moonlight_balance)
-                }
-            };
+            let (addr, balance) = pick_transaction_model(
+                wallet,
+                profile_idx,
+                phoenix_spendable,
+                moonlight_balance,
+            )?;
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_STAKE,
+                "a stake transaction",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
             ProfileOp::Run(Box::new(Command::Stake {
                 address: Some(addr),
                 amt: prompt::request_stake_token_amt(balance)?,
@@ -114,30 +139,47 @@ pub(crate) fn online(
             }))
         }
         MenuItem::Unstake => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
+            let (addr, balance) = pick_transaction_model(
+                wallet,
+                profile_idx,
+                phoenix_spendable,
+                moonlight_balance,
+            )?;
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_STAKE,
+                "an unstake transaction",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
             ProfileOp::Run(Box::new(Command::Unstake {
                 address: Some(addr),
                 gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
                 gas_price: prompt::request_gas_price()?,
             }))
         }
-
         MenuItem::Withdraw => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
+            let (addr, balance) = pick_transaction_model(
+                wallet,
+                profile_idx,
+                phoenix_spendable,
+                moonlight_balance,
+            )?;
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_STAKE,
+                "a stake reward withdrawal transaction",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
             ProfileOp::Run(Box::new(Command::Withdraw {
                 address: Some(addr),
                 gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
@@ -145,14 +187,23 @@ pub(crate) fn online(
             }))
         }
         MenuItem::ContractDeploy => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
+            let (addr, balance) = pick_transaction_model(
+                wallet,
+                profile_idx,
+                phoenix_spendable,
+                moonlight_balance,
+            )?;
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_DEPLOYMENT,
+                "a deploy transaction",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
             ProfileOp::Run(Box::new(Command::ContractDeploy {
                 address: Some(addr),
                 code: prompt::request_contract_code()?,
@@ -165,14 +216,23 @@ pub(crate) fn online(
             }))
         }
         MenuItem::ContractCall => {
-            let addr = match prompt::request_transaction_model()? {
-                prompt::TransactionModel::Shielded => {
-                    wallet.shielded_address(profile_idx)
-                }
-                prompt::TransactionModel::Public => {
-                    wallet.public_address(profile_idx)
-                }
-            }?;
+            let (addr, balance) = pick_transaction_model(
+                wallet,
+                profile_idx,
+                phoenix_spendable,
+                moonlight_balance,
+            )?;
+
+            if check_min_gas_balance(
+                balance,
+                DEFAULT_LIMIT_CALL,
+                "a contract call",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
             ProfileOp::Run(Box::new(Command::ContractCall {
                 address: Some(addr),
                 contract_id: prompt::request_bytes("contract id")?,
@@ -195,18 +255,42 @@ pub(crate) fn online(
             profile_idx: Some(profile_idx),
             reward: false,
         })),
-        MenuItem::Shield => ProfileOp::Run(Box::new(Command::Shield {
-            profile_idx: Some(profile_idx),
-            amt: prompt::request_token_amt("convert", moonlight_balance)?,
-            gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-            gas_price: prompt::request_gas_price()?,
-        })),
-        MenuItem::Unshield => ProfileOp::Run(Box::new(Command::Unshield {
-            profile_idx: Some(profile_idx),
-            amt: prompt::request_token_amt("convert", phoenix_balance)?,
-            gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
-            gas_price: prompt::request_gas_price()?,
-        })),
+        MenuItem::Shield => {
+            if check_min_gas_balance(
+                moonlight_balance,
+                DEFAULT_LIMIT_CALL,
+                "convert DUSK from public to shielded",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
+            ProfileOp::Run(Box::new(Command::Shield {
+                profile_idx: Some(profile_idx),
+                amt: prompt::request_token_amt("convert", moonlight_balance)?,
+                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
+                gas_price: prompt::request_gas_price()?,
+            }))
+        }
+        MenuItem::Unshield => {
+            if check_min_gas_balance(
+                phoenix_spendable,
+                DEFAULT_LIMIT_CALL,
+                "convert DUSK from shielded to public",
+            )
+            .is_err()
+            {
+                return Ok(ProfileOp::Stay);
+            }
+
+            ProfileOp::Run(Box::new(Command::Unshield {
+                profile_idx: Some(profile_idx),
+                amt: prompt::request_token_amt("convert", phoenix_spendable)?,
+                gas_limit: prompt::request_gas_limit(gas::DEFAULT_LIMIT_CALL)?,
+                gas_price: prompt::request_gas_price()?,
+            }))
+        }
         MenuItem::CalculateContractId => {
             ProfileOp::Run(Box::new(Command::CalculateContractId {
                 profile_idx: Some(profile_idx),
@@ -260,4 +344,46 @@ pub(crate) fn offline(
         _ => unreachable!(),
     };
     Ok(res)
+}
+
+/// Prompts the user to select a transaction model (Shielded or Public), and
+/// retrieves the corresponding address and balance for the specific profile
+fn pick_transaction_model(
+    wallet: &Wallet<WalletFile>,
+    profile_idx: u8,
+    phoenix_spendable: Dusk,
+    moonlight_balance: Dusk,
+) -> anyhow::Result<(Address, Dusk)> {
+    match prompt::request_transaction_model()? {
+        prompt::TransactionModel::Shielded => {
+            let addr = wallet.shielded_address(profile_idx)?;
+            Ok((addr, phoenix_spendable))
+        }
+        prompt::TransactionModel::Public => {
+            let addr = wallet.public_address(profile_idx)?;
+            Ok((addr, moonlight_balance))
+        }
+    }
+}
+
+/// Verifies that the user's balance meets the minimum required gas for a given
+/// action
+fn check_min_gas_balance(
+    balance: Dusk,
+    min_required_gas: u64,
+    action: &str,
+) -> anyhow::Result<()> {
+    let min_required_gas: Dusk = min_required_gas.into();
+    if balance < min_required_gas {
+        println!(
+            "Balance too low to cover the minimum gas cost for {}.",
+            action
+        );
+        Err(anyhow::anyhow!(
+            "Balance too low to cover the minimum gas cost for {}.",
+            action
+        ))
+    } else {
+        Ok(())
+    }
 }